Pr 25 merged in 26 & Refactorized
