+++ frontend/src/components/host/CardForm.jsx
@@ -1,4 +1,3 @@
-import React from "react";
 import styles from "@/styles/components/host/CardForm.module.css";
 
 const CardForm = ({ index, card, onCardChange, onDeleteCard }) => {
