--- frontend/src/components/neonCard.jsx
@@ -1,9 +1,0 @@
-import React from "react";
-import '../styles/components/neonCard.css';
-
-const NeonCard = () =>{
-    return(
-        <button className="flip cardStyle"><p className="tit">?</p></button>
-    )
-}
-export default NeonCard;
