+++ frontend/src/components/player/Board/Board.jsx
@@ -1,0 +1,43 @@
+import NeonCard from '@/components/player/Board/Card'
+import styles from "@/styles/components/player/Board.module.css"
+
+
+export default function Board() {
+
+    return (
+        <div className={styles.right}>
+            <NeonCard />
+            <NeonCard />
+            <NeonCard />
+            <NeonCard />
+            <NeonCard />
+            <NeonCard />
+            <NeonCard />
+            <NeonCard />
+            <NeonCard />
+            <NeonCard />
+            <NeonCard />
+            <NeonCard />
+            <NeonCard />
+            <NeonCard />
+            <NeonCard />
+            <NeonCard />
+            <NeonCard />
+            <NeonCard />
+            <NeonCard />
+            <NeonCard />
+            <NeonCard />
+            <NeonCard />
+            <NeonCard />
+            <NeonCard />
+            <NeonCard />
+            <NeonCard />
+            <NeonCard />
+            <NeonCard />
+            <NeonCard />
+            <NeonCard />
+            <NeonCard />
+            <NeonCard />
+        </div>
+    );
+}
