+++ frontend/src/components/player/InfoCard.jsx
