+++ frontend/src/styles/components/host/GameForm.module.css
@@ -1,5 +1,6 @@
 .formContainer {
   margin-top: 20px;
+  text-align: center;
 }
 
 .labelInput {
@@ -36,9 +37,9 @@
   padding: 10px 15px;
   border: none;
   cursor: pointer;
-  margin-left: 120px;
+  margin-right: 20px;
+  margin-left: 20px;
   border-radius: 5px;
-  margin-right: 10px;
   margin-top: 10px;
 }
