+++ frontend/src/styles/components/loading.module.css
@@ -9,7 +9,7 @@
 }
 
 .loader {
-    background: conic-gradient( #161b3d, #d90368);
+    background: conic-gradient( black, #d90368);
     position: relative;
     border-radius: 50%;
     width: 100px;
@@ -24,7 +24,7 @@
     border-radius: 50%;
     width: 80px;
     height: 80px;
-    background-color: #161b3d;
+    background-color: black;
     top: 10%;
     left: 10%;
 }
