+++ frontend/src/styles/components/player/AsideTeamInfo.module.css
@@ -3,6 +3,7 @@
     height: 80vh;
     margin: 3vw;
     position: absolute;
+    
 }
 .titleContainer{
     width: 100%;
@@ -16,6 +17,9 @@
     font-style: normal;
     color: white;
     margin-left: 1vw;
+    
+    margin: 0;
+    padding: 0;
 }
 .tit{
     font-family: "Londrina Solid", sans-serif;
@@ -48,20 +52,11 @@
     margin-top: 1vh;
     margin-bottom: 1vh;
 }
-.playerScore{
-    width: 100%;
-    height: 50vh;
-    display: flex;
-    flex-wrap: wrap;
-}
+
 .line{
     width: 18vw;
     height: 45vh;
     border-right: 1px solid white;
     position: absolute;
 
-}
-hr{
-    margin: 1vw;
-    background-color: rgb(255, 255, 255, 0.5);
 }
