--- frontend/src/styles/components/player/Board.css
@@ -1,56 +1,0 @@
-body {
-  height: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.card-container {
-  display: grid;
-  grid-template-columns: repeat(4, 1fr);
-  gap: 1rem;
-  background-color: #161b3d;
-  padding: 1rem;
-  perspective: 700px;
-}
-
-.card {
-  background-color: white;
-  height: 100px;
-  width: 100px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  position: relative;
-  transform-style: preserve-3d;
-  transition: 0.3s ease;
-
-  border: #d90368 solid 2px;
-  border-radius: 8px;
-}
-
-.animalName {
-  color: black;
-  object-fit: cover;
-}
-
-.back {
-  background: url("https://static.vecteezy.com/system/resources/thumbnails/007/126/739/small/question-mark-icon-free-vector.jpg")
-    center/cover;
-  background-color: white;
-  height: 100%;
-  width: 100%;
-  position: absolute;
-  top: 0;
-  backface-visibility: hidden;
-  cursor: pointer;
-  border-radius: 8px;
-}
-
-.card.open {
-  transform: rotateY(180deg);
-}
-
-.score {
-  text-align: center;
-}
