+++ frontend/src/styles/components/player/Board.module.css
@@ -1,0 +1,10 @@
+.right{
+  width: 65vw;
+  height: 90vh;
+  position: absolute;
+  right:3vw;
+  top: 5vh;
+  display: flex;
+  flex-wrap: wrap;
+  align-content: flex-start;
+}
