+++ frontend/src/styles/components/player/Card.module.css
