+++ frontend/src/styles/components/player/CocoDeTeamFortress.css
@@ -1,0 +1,4 @@
+hr{
+    margin: 1vw;
+    background-color: rgb(255, 255, 255, 0.5);
+}
