+++ frontend/src/styles/pages/create.module.css
@@ -1,7 +1,8 @@
 .container {
-  max-width: 600px;
   margin: 0 auto;
   padding: 20px;
+  height: 95vh;
+  overflow-y: scroll;
 }
 
 .title {
