--- frontend/src/styles/pages/edicionJuego.css
@@ -1,38 +1,0 @@
-* {
-  margin: 0;
-  padding: 0;
-  font-family: "Montserrat", sans-serif;
-  font-optical-sizing: auto;
-  font-weight: 400;
-  font-style: normal;
-}
-
-html,
-body {
-  width: 100%;
-  height: 100%;
-}
-
-body {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-/* main {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-} */
-
-form {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-}
-
-form .input {
-  margin-top: 1.5vh;
-  margin-bottom: 3vh;
-}
