--- frontend/src/styles/pages/formularioIngreso.css
@@ -1,70 +1,0 @@
-* {
-  margin: 0;
-  padding: 0;
-  font-family: "Montserrat", sans-serif;
-  font-optical-sizing: auto;
-  font-weight: 400;
-  font-style: normal;
-}
-
-html,
-body {
-  width: 100%;
-  height: 100%;
-}
-
-body {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background-color: #161b3d;
-}
-
-
-.box {
-  background-color: white;
-  padding: 1.8rem 1rem;
-  border-radius: 20px;
-  border: #d90368 solid 2px;
-  color: black;
-}
-
-#ingresar-form {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-}
-
-#ingresar-form .input {
-  font-size: 1.1rem;
-  padding: 0 0.1rem;
-  padding-bottom: 0.1rem;
-  margin-top: 1rem;
-  margin-bottom: 2rem;
-  border: none;
-  outline: none;
-  border-bottom: #d90368 solid 2px;
-}
-
-#ingresar-form label {
-  font-size: 1.1rem;
-}
-
-
-#submit-btn {
-  margin-top: 5vh;
-  padding: 0.4rem 0.6rem;
-  border-radius: 16px;
-  transition: all 0.3s ease-in-out;
-  cursor: pointer;
-  background-color: #d90368;
-  border: #d90368 solid 2px;
-  color: white;
-}
-
-#submit-btn:hover {
-  transform: scale(1.05);
-  background-color: white;
-  color: black;
-}
