--- frontend/src/styles/pages/gameView.css
@@ -1,21 +1,0 @@
-*{
-    padding: 0%;
-    margin: 0%;
-    font-family: "Londrina Solid", sans-serif;
-}
-body{
-    width: 100%;
-    height: 100%;
-    background-color: black;
-    overflow: hidden;
-}
-.right{
-    width: 65vw;
-    height: 90vh;
-    position: absolute;
-    right:3vw;
-    top: 5vh;
-    display: flex;
-    flex-wrap: wrap;
-    align-content: flex-start;
-}
