--- frontend/src/styles/pages/teachersRoom.css
@@ -1,95 +1,0 @@
-*{
-    margin: 0%;
-    padding: 0%;
-}
-html, body{
-    background-color: #1E1E1E;
-}
-
-.App{
-    min-height: 100vh;
-    display: flex;
-    flex-wrap: wrap;
-    align-items: center;
-    justify-content: center;
-    text-align: center;
-}
-.contenedor{
-    width: 50vw;
-    height: 70vh;
-    border: 2px solid white;
-    box-shadow: 0 0 10px 5px rgb(106, 213, 255);
-    align-items: center;
-}
-.id-sala{
-    width: 20%;
-    height: 5vh;
-    border: white;
-    box-shadow: 0 0 10px 3px rgb(106, 213, 255);
-    border: 2px solid white;
-    float: left;
-    margin-top: 5vh;
-    margin-left: 5%;
-}
-.nombre-sala{
-    width: 65%;
-    height: 5vh;
-    border: white;
-    box-shadow: 0 0 10px 3px rgb(106, 213, 255);
-    border: 2px solid white;
-    float: left;
-    margin-top: 5vh;
-    margin-left: 5%;
-}
-#nombre-sala-r{
-    color: white;
-}
-#nuevo-nombre{
-    width: 65%;
-    height: 5vh;
-    border: white;
-    box-shadow: 0 0 10px 3px rgb(106, 213, 255);
-    border: 2px solid white;
-    float: left;
-    margin-top: 5vh;
-    margin-left: 5%;
-}
-.bot-nombre{
-    width: 20%;
-    height: 5vh;
-    background-color: #1E1E1E;
-    color: white;
-    border: 2px solid white;
-    float: left;
-    margin-top: 5vh;
-    margin-left: 5%;
-}
-.agregar-mazo{
-    width: 20%;
-    height: 5vh;
-    background-color: #1E1E1E;
-    color: white;
-    border: 2px solid white;
-    float: left;
-    margin-top: 5vh;
-    margin-left: 5%;
-}
-.jugadores-conectados{
-    width: 90%;
-    height:23vh;
-    box-shadow: 0 0 10px 3px rgb(106, 213, 255);
-    border: 2px solid white;
-    float: left;
-    margin-top: 5vh;
-    margin-left: 5%;
-}
-.jugar{
-    width: 20%;
-    height: 5vh;
-    background-color: #1E1E1E;
-    color: white;
-    border: 2px solid white;
-    float: left;
-    margin-top: 3vh;
-    margin-left: 40%;
-}
